PAGOPA-1837 refactoring
