+++ src/test/java/it/gov/pagopa/wispconverter/endpoint/RptTest.java
@@ -1,6 +1,6 @@
-package it.gov.pagopa.wispconverter;
+package it.gov.pagopa.wispconverter.endpoint;
 
-import static it.gov.pagopa.wispconverter.ConstantsTestHelper.REDIRECT_PATH;
+import static it.gov.pagopa.wispconverter.utils.ConstantsTestHelper.REDIRECT_PATH;
 import static it.gov.pagopa.wispconverter.utils.TestUtils.*;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -11,6 +11,7 @@ import it.gov.pagopa.gen.wispconverter.client.cache.model.StationCreditorInstitu
 import it.gov.pagopa.gen.wispconverter.client.iuvgenerator.model.IUVGenerationResponseDto;
 import it.gov.pagopa.gen.wispconverter.client.gpd.model.MultiplePaymentPositionModelDto;
 import it.gov.pagopa.gen.wispconverter.client.gpd.model.PaymentPositionModelDto;
+import it.gov.pagopa.wispconverter.Application;
 import it.gov.pagopa.wispconverter.repository.CacheRepository;
 import it.gov.pagopa.wispconverter.repository.RPTRequestRepository;
 import it.gov.pagopa.wispconverter.repository.RTRequestRepository;
