+++ src/test/java/it/gov/pagopa/wispconverter/utility/XmlUtilTest.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.wispconverter;
+package it.gov.pagopa.wispconverter.utility;
 
 import it.gov.pagopa.wispconverter.util.XmlUtil;
 import org.junit.Test;
